extracting database handler into its own files
